rag/requirements.txt
pathway[all]==0.20.0
fastapi
uvicorn
openai
python-dotenv
gunicorn